+++ crates/driver_usb/src/host/mod.rs
@@ -100,7 +100,10 @@ where
                 trace!("request transfer!");
                 self.control_transfer(request.device_slot_id, control)
             }
-            usb::urb::RequestedOperation::Bulk => todo!(),
+            usb::urb::RequestedOperation::Bulk(BulkTransfer) => self
+                .controller
+                .lock()
+                .bulk_transfer(request.device_slot_id, BulkTransfer),
             usb::urb::RequestedOperation::Interrupt(interrupt_transfer) => self
                 .controller
                 .lock()